@@ -1,6 +1,5 @@
-import { ParsedUrlQuery } from "querystring"
-
-import { useRouter } from "next/router"
+"use client"
+import { useRouter, useSearchParams } from "next/navigation"
 import React, { useEffect } from "react"
 import Container from "react-bootstrap/Container"
 import Image from "react-bootstrap/Image"
@@ -8,12 +7,12 @@ import Image from "react-bootstrap/Image"
 import CurrencyInput, { formatValue } from "react-currency-input-field"
 
 import useRealtimePrice from "../../lib/use-realtime-price"
-import { ACTION_TYPE, ACTIONS } from "../../pages/_reducer"
+import { ACTION_TYPE, ACTIONS } from "../../app/reducer"
 import {
   formatOperand,
-  parseDisplayCurrency,
   safeAmount,
   getLocaleConfig,
+  extractSearchParams,
 } from "../../utils/utils"
 import Memo from "../Memo"
 
@@ -26,14 +25,18 @@ import styles from "./parse-payment.module.css"
 import ReceiveInvoice from "./Receive-Invoice"
 
 function isRunningStandalone() {
+  if (typeof window === "undefined") {
+    return false
+  }
   return window.matchMedia("(display-mode: standalone)").matches
 }
 
 interface Props {
-  defaultWalletCurrency?: string
-  walletId?: string
+  defaultWalletCurrency: string
+  walletId: string
   dispatch: React.Dispatch<ACTION_TYPE>
   state: React.ComponentState
+  username: string
 }
 
 interface UpdateAmount {
@@ -56,10 +59,18 @@ const defaultCurrencyMetadata: Currency = {
   __typename: "Currency",
 }
 
-function ParsePayment({ defaultWalletCurrency, walletId, dispatch, state }: Props) {
+function ParsePayment({
+  defaultWalletCurrency,
+  walletId,
+  dispatch,
+  state,
+  username,
+}: Props) {
   const router = useRouter()
-  const { username, amount, sats, unit, memo, currency } = router.query
-  const { display } = parseDisplayCurrency(router.query)
+  const searchParams = useSearchParams()
+  const { amount, sats, unit, memo } = extractSearchParams(searchParams)
+
+  const display = searchParams?.get("display") ?? localStorage.getItem("display") ?? "USD"
   const { currencyToSats, satsToCurrency, hasLoaded } = useRealtimePrice(display)
   const { currencyList } = useDisplayCurrency()
   const [valueInFiat, setValueInFiat] = React.useState(0)
@@ -70,24 +81,8 @@ function ParsePayment({ defaultWalletCurrency, walletId, dispatch, state }: Prop
     defaultCurrencyMetadata,
   )
   const [numOfChanges, setNumOfChanges] = React.useState(0)
-
+  const language = typeof navigator !== "undefined" ? navigator?.language : "en"
   const prevUnit = React.useRef(AmountUnit.Cent)
-
-  if (!currency) {
-    const queryString = window.location.search
-    const searchParams = new URLSearchParams(queryString)
-    searchParams.set("currency", defaultWalletCurrency ?? "BTC")
-    const newQueryString = searchParams.toString()
-    window.history.pushState(null, "", "?" + newQueryString)
-  }
-
-  if (!amount) {
-    const queryString = window.location.search
-    const searchParams = new URLSearchParams(queryString)
-    searchParams.set("amount", "0")
-    const newQueryString = searchParams.toString()
-    window.history.pushState(null, "", "?" + newQueryString)
-  }
 
   // onload
   // set all query params on first load, even if they are not passed
@@ -96,10 +91,10 @@ function ParsePayment({ defaultWalletCurrency, walletId, dispatch, state }: Prop
     const initialAmount = safeAmount(amount).toString()
     const initialSats = safeAmount(sats).toString()
     const initialDisplay = display ?? localStorage.getItem("display") ?? "USD"
-    const initialUsername = router.query.username
-    const initialQuery = { ...router.query }
+    const initialUsername = username
+    const initialQuery = extractSearchParams(searchParams)
     delete initialQuery?.currency
-    const newQuery: ParsedUrlQuery = {
+    const newQuery = {
       amount: initialAmount,
       sats: initialSats,
       unit: initialUnit,
@@ -108,21 +103,23 @@ function ParsePayment({ defaultWalletCurrency, walletId, dispatch, state }: Prop
       username: initialUsername,
     }
     if (initialQuery !== newQuery) {
-      router.push(
-        {
-          pathname: `${username}`,
-          query: {
-            amount: initialAmount,
-            sats: initialSats,
-            unit: initialUnit,
-            memo: memo ?? "",
-            display: initialDisplay,
-          },
-        },
-        undefined,
-        { shallow: true },
-      )
+      const params = new URLSearchParams({
+        amount: initialAmount,
+        sats: initialSats,
+        unit: initialUnit,
+        memo: memo ?? "",
+        display: initialDisplay,
+        currency: defaultWalletCurrency,
+      })
+      const newUrl = new URL(window.location.toString())
+      newUrl.pathname = `/${username}`
+      newUrl.search = params.toString()
+
+      router.replace(newUrl.toString(), {
+        scroll: true,
+      })
     }
+    // this only runs once
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 
@@ -136,7 +133,7 @@ function ParsePayment({ defaultWalletCurrency, walletId, dispatch, state }: Prop
       } else if (sats) {
         return {
           shouldUpdate: true,
-          value: sats.toString(),
+          value: sats?.toString(),
         }
       }
     } else {
@@ -152,21 +149,19 @@ function ParsePayment({ defaultWalletCurrency, walletId, dispatch, state }: Prop
   const toggleCurrency = () => {
     const newUnit = unit === AmountUnit.Sat ? AmountUnit.Cent : AmountUnit.Sat
     prevUnit.current = (unit as AmountUnit) || AmountUnit.Cent
-    router.push(
-      {
-        pathname: `${username}`,
-        query: {
-          currency: defaultWalletCurrency,
-          unit: newUnit,
-          memo,
-          display,
-          amount,
-          sats,
-        },
-      },
-      undefined,
-      { shallow: true },
-    )
+    const params = new URLSearchParams({
+      currency: defaultWalletCurrency,
+      unit: newUnit,
+      memo,
+      display,
+      amount,
+      sats,
+    })
+
+    const newUrl = new URL(window.location.toString())
+    newUrl.pathname = `/${username}`
+    newUrl.search = params.toString()
+    router.replace(newUrl.toString())
   }
 
   // Update Params From Current Amount
@@ -190,9 +185,10 @@ function ParsePayment({ defaultWalletCurrency, walletId, dispatch, state }: Prop
           : safeAmt.toFixed(currencyMetadata.fractionDigits)
     }
     if (isNaN(Number(amt))) return
+
     const formattedValue = formatValue({
       value: amt,
-      intlConfig: { locale: navigator.language, currency: display },
+      intlConfig: { locale: language, currency: display },
     })
     localStorage.setItem("formattedFiatValue", formattedValue)
     setValueInFiat(amt)
@@ -216,15 +212,13 @@ function ParsePayment({ defaultWalletCurrency, walletId, dispatch, state }: Prop
       memo,
       display,
     }
-    if (router.query !== newQuery && !skipRouterPush) {
-      router.push(
-        {
-          pathname: `${username}`,
-          query: newQuery,
-        },
-        undefined,
-        { shallow: true },
-      )
+
+    const initialQuery = extractSearchParams(searchParams)
+    if (initialQuery !== newQuery && !skipRouterPush) {
+      const newUrl = new URL(window.location.toString())
+      newUrl.pathname = `/${username}`
+      newUrl.search = new URLSearchParams(newQuery).toString()
+      router.replace(newUrl.toString())
     }
   }
   // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -323,7 +317,7 @@ function ParsePayment({ defaultWalletCurrency, walletId, dispatch, state }: Prop
               fontWeight: 600,
             }}
             value={!amount ? 0 : valueInFiat}
-            intlConfig={{ locale: navigator.language, currency: display }}
+            intlConfig={{ locale: language, currency: display }}
             readOnly={true}
           />
         </div>
@@ -388,8 +382,7 @@ function ParsePayment({ defaultWalletCurrency, walletId, dispatch, state }: Prop
               dispatch={dispatch}
               disabled={unit === AmountUnit.Sat}
               displayValue={
-                getLocaleConfig({ locale: navigator.language, currency: display })
-                  .decimalSeparator
+                getLocaleConfig({ locale: language, currency: display }).decimalSeparator
               }
             />
           ) : (
